fix: move share utilities to a new file and remove unused imageUtils
